requirements.txt
streamlit==1.17.0
PyMuPDF==1.20.0
pytesseract==0.3.8
pdf2image==1.16.0
google-generativeai==0.2.0  # Ensure this matches the latest version or your required version
Pillow==9.4.0
python-dotenv==1.0.0
